pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-log-remediator"
version = "1.0.0"
requires-python = ">=3.10"

dependencies = [
  "python-dotenv",
  "aiofiles",
  "sentence-transformers",
  "chromadb",
  "gitpython",
  "PyGithub",
  "groq",
  "requests",
  "langchain",
  "langchain-community",
  "langchain-groq"
]

[tool.setuptools]
packages = ["app"]